requirements.txt
pandas
matplotlib.pyplot
seaborn
scikitlearn
numpy